requirements.txt
tensorflow==2.12.0 
numpy 
pandas 
pillow 
scikit-learn 
matplotlib 
streamlit 
rasterio 
tqdm 
